requirements.txt
Flask==3.0.0
flask-cors==4.0.0
google-generativeai==0.8.3
Pillow==10.4.0
pandas==2.2.0
numpy==1.26.4
scikit-learn==1.4.0
joblib==1.3.2
python-dotenv==1.0.0
Werkzeug==3.0.1
requests==2.31.0
gunicorn==21.2.0
